Adds an abstraction to the user agent time series computer.
This change reduces code duplication by extracting the common used functionality.

Fixes: OX-9081
